Externalize the common logic for controllers
